Bug fixes, QoL improvements
